# Memprof: optimize random sampling