This PR includes two optimizations to improve the performance of random sampling in Memprof:
1) Instead of using the stdlib logf function for computing logarithms, we use a faster polynomial-based approximation.

This is using the same ideas I have already used in another project : https://github.com/jhjourdan/SIMD-math-prims/. The code is portable enough (the least portable part is that it depends on the relative endianness of float and uint32_t, which I think is always the same in OCaml architecture).

2) We generate samples by batches so that we let compilers optimize the generation loops using SIMD instructions when possible.

I tested (using godbolt)  on several versions of CLang (>4.0.0), GCC (> 4.7), and they all produce vectorized code. Of course, other compilers (e.g., MSVC, ICC) will not perform these optimizations, but the code is then still more efficient than the current one.

I benchmarked with a micro-benchmark adapted from @stedolan's #8731 

```ocaml
let samples = ref 0

let rec recur n f = match n with 0 -> f () | n -> recur (n-1) f + 1

let () =
  Gc.Memprof.start
    ~sampling_rate:0.1
    ~callstack_size:10
    { Gc.Memprof.null_tracker with
      alloc_minor = fun _ -> incr samples; None };

  for i = 1 to 10_000_000 do
    let d = recur 20 (fun () ->
      Array.make 20 0 |> ignore; ref 42 |> Sys.opaque_identity |> ignore; 0) in
    assert (d = 20)
  done;

  Printf.printf "%d\n" !samples
```

Without these optimizations, on my machine, this code takes about 3.6 seconds. With this new code but without vectorization, it takes 3.4 seconds, and with vectorization, it takes 3.1 seconds. Given that the runtime of this program is about 0.7 seconds when statmemprof is disabled, this mean that this patch improves by about 17% the overhead of statmemprof.

This is an improvement, but I am still not completely sure this is worth it, given, in particular, that it requires some changes in `configure` to detect the specific attributes we need to pass to GCC to tell it to optimize. What do people think?

Note that @stedolan proposed in #8731 and optimization that should improve the performances of `caml_next_frame_descriptor`, which is now the main bottleneck.