@gasche 2020-04-18
> Note that @stedolan proposed in #8731 and optimization that should improve the performances of `caml_next_frame_descriptor`, which is now the main bottleneck.

What part of #8371 are you referring to? There were several things in this PR and a complex discussion, but regarding backtrace construction optimization I understand that there were two changes (two commits):
1. one to avoid rescanning the stack twice that was already merged as #9279 (so I guess it is active in your benchmark results), and
2. another that [uses registers for locals ](https://github.com/ocaml/ocaml/pull/8731/commits/c308c43ae71fe08bf20dae34dc1c8b1293d65d20) that, if I understand, was never integrated

Are you talking about this second change, or another optimization?