@jhjourdan 2020-04-18
> Are you talking about this second change, or another optimization?

I am talking about that second change, yes.