@gasche 2020-04-18
Thanks. (@stedolan mentioned a 10-15% difference, so that would shave maybe 0.3-0.4 seconds off your benchmark.)