@gadmm 2020-04-19
> This is an improvement, but I am still not completely sure this is worth it

I like the idea because I use it with callstack_size = 0, and I often do very little work in the callbacks, in which case the benefit is even greater, and the higher I can chose a sampling rate without affecting performance the better.

The software you mention (https://github.com/jhjourdan/SIMD-math-prims/) says that there is a loss of precision. Is this relevant here? If so can you give us an idea of how this skews the distribution? (e.g. the impact on the number of samples expected after N allocations for N very large).