@xavierleroy 2020-04-20
I'm reminded of the famous [Quake III inverse square root function](https://medium.com/hard-mode/the-legendary-fast-inverse-square-root-e51fee3b49d9), but we need fewer comments and more profanity to rise to the Quake III level :-)

I confirm that the code vectorizes perfectly with SSE, AVX2, and all the way to AVX512.  Impressive!

As to accuracy, the code claims 10^-6 accuracy/ I can't say whether it's enough or not, but note that single-precision FP gives 10^-7 accuracy at most.  So, if single FP is good enough, the proposed code is probably good enough.