@gadmm 2020-04-20
> As to accuracy, the code claims 10^-6 accuracy/ I can't say whether it's enough or not, but note that single-precision FP gives 10^-7 accuracy at most. So, if single FP is good enough, the proposed code is probably good enough.

Maybe, maybe not, and maybe it was already too low before and my use of memprof reminds of a famous bug in a weapon that had to be rebooted every eight hours to remain accurate. Is it possible to convert the figure into an error range for the sampling rate? Then I can run some calculation to tell you if it has an impact.

If it does, at worst I guess it is possible to switch to an even greater precision for very low values of the sampling rate, where the performance loss should not be noticeable.