@jhjourdan 2020-04-20
> I confirm that the code vectorizes perfectly with SSE, AVX2, and all the way to AVX512. Impressive!

Don't expect too much, though: compilers will not do so, except if they are given the right e.g., -march options, which won't happen with OCaml's Makefile.

> As to accuracy, the code claims 10^-6 accuracy/ I can't say whether it's enough or not, but note that single-precision FP gives 10^-7 accuracy at most. So, if single FP is good enough, the proposed code is probably good enough.

10^-6 was too optimistic. The actual precision is about 1.5e-5. But still enough for what we are doing.