@gadmm 2020-04-20
> But still enough for what we are doing.

Thanks for the precisions.