@jhjourdan 2020-04-20
About accuracy: I ran some tests in order to determine the actual impact on sampling rate. It turns out that there are only 2^32 values that can be uniformly generated by the Mersenne Twister. And it turns out that, thanks to vectorization, it is particularly easy to compute the actual sampling rate when requesting a given sampling rate by averaging over all of the 2^32 values.

So, here are the results:

```
lambda = 0.99 error =  6.57301e-08 words = 2.29143e+12
lambda = 0.9 error =  1.33628e-07 words = 5.04016e+12
lambda = 0.2 error =  9.36631e-08 words = 1.82383e+13
lambda = 0.1 error =  6.6075e-08 words = 2.06143e+13
lambda = 0.03 error =  3.19408e-08 words = 2.85234e+13
lambda = 0.01 error =  1.0935e-08 words = 8.27931e+13
lambda = 0.003 error =  3.25604e-09 words = 2.82121e+14
lambda = 0.001 error =  1.15338e-09 words = 7.50972e+14
lambda = 0.0003 error =  3.34931e-10 words = 2.6735e+15
lambda = 0.0001 error =  1.13561e-10 words = 7.75351e+15
lambda = 3e-05 error =  3.28782e-11 words = 2.77518e+16
lambda = 1e-05 error =  1.13485e-11 words = 7.76465e+16
lambda = 3e-06 error =  3.49937e-12 words = 2.44986e+17
lambda = 1e-06 error =  1.13515e-12 words = 7.76062e+17
lambda = 3e-07 error =  3.32747e-13 words = 2.70952e+18
lambda = 1e-07 error =  1.0686e-13 words = 8.75735e+18
lambda = 3e-08 error =  3.20678e-14 words = 2.9173e+19
lambda = 1e-08 error =  1.09474e-14 words = 8.34409e+19
lambda = 3e-09 error =  3.37326e-15 words = 2.63647e+20
lambda = 1e-09 error =  1.09219e-15 words = 8.38314e+20
```

This means that, for example, with lambda = 0.99, one would need about to allocate about 2.29143e+12 words in order to actually be able to observe the bias. Hence, the loss in precision only has negligible impact on the behavior of statmemprof.

Of course, this reasoning rely on the actual uniformity of the Mersenne Twister.