@jhjourdan 2020-04-20
> If it does, at worst I guess it is possible to switch to an even greater precision for very low values of the sampling rate, where the performance loss should not be noticeable.

We could do that, but low values of the sampling rate are values for which the loss in precision is even less noticeable, since there are even less samplings...