@gadmm 2020-04-20
Thanks a lot for looking for those numbers (I do not understand everything, but it looks fun).