@stedolan 2020-04-22
I'm happy to give the maths a read. (I actually wrote one of these a while ago but never got around to making a PR, so it will be fun to compare)