@stedolan 2020-04-28
Looks good to me! There are a few more games you can play here if you like, but moving from libm's `logf` to a polynomial approximation is the most important part.

There have in the past existed machines with different endianness for floats and integers. Are we confident these are all gone?

The polynomial is accurate to about 10^-5 on [1,2], which seems plenty. In this application, we don't really care about accuracy of a single sample, but rather about accuracy of a sum of samples as this is what determines the drift in sample rate. In other words, we (unusually) care about the mean error, rather than the max absolute error. In this respect, the polynomial approximation is even better, with average error about 10^-6 on [1,2].

If you want to optimise further, there are much smaller and faster RNGs than Mersenne Twister. I'd suggest either [PCG](https://www.pcg-random.org/download.html) or [xoroshiro128+](http://prng.di.unimi.it/xoshiro128plus.c). (The former is better in ways that are unlikely to matter much here, the latter will be very fast even on 32-bit systems).